requirements.txt
freetype_py
lz4
moderngl
numpy==1.23
pymeshio
PyQt5
PyQt5_sip
pyrr
transformations
zstandard
